Read Notes!
Changed default models out for presents - will be unable to draw CTF flags for now.

Created new internal array of "BKObjectives" objects, to not compete with the traditional array.

Wrote new routines for creating and displaying objects inside that array.

Presents are created and displayed when hit. Need to add collision